backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-mail==0.10.0
python-dotenv==1.0.1
werkzeug==3.0.1
apscheduler==3.10.4
pymysql==1.1.0
pywebpush==1.14.0
cryptography==42.0.5
